+++ examples/docs/cli_config_inline.py
@@ -10,10 +10,11 @@ class ExecutorConfig(yuio.config.Config):
 
 @yuio.app.app
 def main(
+    #: File to process.
+    input: pathlib.Path,
+    /,
     #: Executor options.
     executor_config: ExecutorConfig = yuio.app.inline(),
-    #: File to process.
-    input: pathlib.Path = yuio.app.positional(),
 ):
     ...