+++ examples/docs/hw_flags.py
@@ -4,8 +4,10 @@ import yuio.io
 
 @yuio.app.app()
 def main(
-    greeting: str = yuio.app.field(default="world", flags=["-g", "--greeting"]),  # [2]_
-    output: pathlib.Path | None = yuio.app.positional(default=None),  # [3]_
+    output: pathlib.Path | None = None,  # [2]_
+    /,  # [3]_
+    *,
+    greeting: str = yuio.app.field(default="world", flags=["-g", "--greeting"]),  # [4]_
 ):
     if output:
         output.write_text(f"Hello, {greeting}!\n")